requirements.txt
absl-py==2.1.0
cached-property @ file:///home/conda/feedstock_root/build_artifacts/cached_property_1615209429212/work
cachetools==5.5.0
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
easydict @ file:///home/conda/feedstock_root/build_artifacts/easydict_1731199109457/work
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1714285159399/work
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.55.1
fsspec==2024.10.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.68.1
h5py @ file:///home/conda/feedstock_root/build_artifacts/h5py_1717664826778/work
humanfriendly==10.0
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
loguru @ file:///home/conda/feedstock_root/build_artifacts/loguru_1695547325717/work
Markdown==3.7
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1686175045316/work
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1704317613764/work
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
onnx==1.17.0
onnx-simplifier==0.4.36
onnxoptimizer==0.3.13
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
onnxsim==0.4.36
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1733203243479/work
pandas==2.0.3
pillow==10.4.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.4.1
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1723018227672/work
requests-oauthlib==2.0.0
rich @ file:///home/conda/feedstock_root/build_artifacts/rich_1730592237829/work/dist
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sympy==1.13.3
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==0.5.4
torch==2.1.0
torch-tb-profiler==0.4.3
torchvision==0.16.0
tqdm==4.67.1
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.6
zipp==3.20.2